devops_agent/requirements.txt
requests>=2.28.0
pyyaml>=6.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.0.0 